@@ -15,20 +15,23 @@
     "electron:serve": "wait-on ./dist/electron-angular/index.html && npm run electron:watch",
     "ng:build": "ng build --watch=true --poll 1000 ",
     "electron:tsc": "tsc -w main.ts --lib es2015,es2016,es2017,esnext,dom",
-    "electron:run": "wait-on main.js && electron . --serve --console --development",
+    "electron:run": "wait-on main.js && electron . --serve --console --development --remote-debugging-port=9222",
     "electron:exit": "taskkill /IM electron.exe /F",
     "electron:watch": "concurrently --kill-others \"npm run electron:tsc \"  \"npm run electron:run\""
+  },
+  "browser": {
+    "fs": false,
+    "path": false,
+    "os": false,
+    "net": false,
+    "tls": false,
+    "zlib": false
   },
   "build": {
     "asar": true,
     "productName": "Decentral Mediaserver",
     "appId": "com.electron.decentral-mediaserver",
     "copyright": "Mabenan",
-    "publish": {
-      "provider": "s3",
-      "bucket": "com.electron.decentral-mediaserver",
-      "path": "/updates"
-    },
     "files": [
       "./main.js",
       "dist"
@@ -50,12 +53,7 @@
       "target": [
         "nsis"
       ],
-      "icon": "./src/assets/images/icon.ico",
-      "publish": {
-        "provider": "s3",
-        "bucket": "com.electron.decentral-mediaserver",
-        "path": "/updates"
-      }
+      "icon": "./src/assets/images/icon.ico"
     },
     "linux": {
       "category": "Music",
@@ -74,25 +72,42 @@
   "dependencies": {
     "@angular/animations": "^7.1.3",
     "@angular/cdk": "^7.1.1",
-    "@angular/material": "^7.1.1",
+    "@angular/material": "^7.3.7",
+    "@angular/platform-server": "^8.1.3",
+    "@nguniversal/module-map-ngfactory-loader": "^8.1.1",
+    "@types/async": "^3.0.1",
+    "@types/bootstrap": "^4.2.1",
+    "@types/ftp": "^0.3.29",
+    "@types/jquery": "^3.3.30",
+    "async": "^3.1.0",
+    "bootstrap": "^4.2.1",
+    "buffer": "^5.2.1",
+    "chart.js": "^2.5.0",
     "core-js": "^2.5.4",
+    "events": "^3.0.0",
+    "express": "^4.17.1",
+    "ftp": "^0.3.10",
+    "jquery": "^3.4.1",
+    "popper.js": "^1.15.0",
     "rxjs": "^6.0.0",
+    "stream": "0.0.2",
+    "timers": "^0.1.1",
     "tslib": "^1.9.0",
     "yargs": "^12.0.5",
-    "zone.js": "^0.8.26"
+    "zone.js": "^0.8.29"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^0.13.1",
     "@angular/cli": "~7.1.4",
-    "@angular/common": "^7.1.3",
-    "@angular/compiler": "^7.1.3",
-    "@angular/compiler-cli": "^7.1.3",
-    "@angular/core": "^7.1.3",
+    "@angular/common": "^7.2.15",
+    "@angular/compiler": "^7.2.15",
+    "@angular/compiler-cli": "^7.2.15",
+    "@angular/core": "^7.2.15",
     "@angular/forms": "^7.1.3",
     "@angular/http": "^7.1.3",
     "@angular/language-service": "^7.1.3",
-    "@angular/platform-browser": "^7.1.3",
-    "@angular/platform-browser-dynamic": "^7.1.3",
+    "@angular/platform-browser": "^7.2.15",
+    "@angular/platform-browser-dynamic": "^7.2.15",
     "@angular/router": "^7.1.3",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
@@ -100,20 +115,22 @@
     "codelyzer": "~4.2.1",
     "concurrently": "^3.6.1",
     "electron": "^3.1.3",
-    "electron-builder": "^20.34.0",
+    "electron-builder": "^20.44.4",
     "electron-reload": "^1.4.0",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^3.1.1",
+    "karma": "^4.2.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.0",
+    "karma-coverage-istanbul-reporter": "^2.0.6",
     "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
     "npm-run-all": "^4.1.5",
-    "protractor": "^5.4.1",
+    "protractor": "^5.4.2",
+    "ts-loader": "^6.0.4",
     "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "^3.1.6",
-    "wait-on": "^2.1.0"
+    "typescript": "^3.2.4",
+    "wait-on": "^2.1.0",
+    "webpack-cli": "^3.3.6"
   }
 }
