@@ -43,7 +43,7 @@
 
 /** Evergreen browsers require these. **/
 // Used for reflect-metadata in JIT. If you use AOT (and only Angular decorators), you can remove.
-import 'core-js/es7/reflect';
+import "core-js/es7/reflect";
 
 
 /**
@@ -71,9 +71,9 @@ import 'core-js/es7/reflect';
 /***************************************************************************************************
  * Zone JS is required by default for Angular itself.
  */
-import 'zone.js/dist/zone';  // Included with Angular CLI.
-
-
+import "zone.js/dist/zone";  // Included with Angular CLI.
+window["$"] = window["jQuery"] = require("jquery"); // not sure if you need this at all
+window["Bootstrap"] = require("bootstrap");
 
 /***************************************************************************************************
  * APPLICATION IMPORTS
